requirements.txt
certifi==2024.6.2
charset-normalizer==3.3.2
dnspython==2.6.1
idna==3.7
numpy==2.0.0
pymongo==4.8.0
python-dotenv==1.0.1
requests==2.32.3
urllib3==1.26.7
xmltodict==0.13.0
